requirements.txt
Flask==2.3.2
flask-login
flask-bcrypt
flask-mongoengine
gunicorn
torch==1.13.1
numpy
opencv-python-headless
Pillow
gdown
fpdf
